+++ packages/admin-api-client/src/types/Common.ts
@@ -1,0 +1,2 @@
+export type EntityId = string
+export type AnyObject = Record<string, unknown>